# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EduFlow API",
        "description": "API de gestion pédagogique",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Serveur local"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Inscription d'un utilisateur",
                "operationId": "08136088c4862c313b35d6518bab8d3a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "role"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Alice"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "alice@test.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "student",
                                            "teacher"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Inscription réussie"
                    },
                    "422": {
                        "description": "Validation échouée"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Connexion",
                "operationId": "99d66635c4992aeaa6aa44ff653d0563",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "alice@test.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie"
                    },
                    "401": {
                        "description": "Identifiants incorrects"
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Déconnexion",
                "operationId": "6abf3b64a4bc7838d56346f05a5153af",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/reset-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Réinitialisation du mot de passe",
                "operationId": "b8d03d5b689dcb65d3771f62d4f90a24",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "alice@test.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lien envoyé"
                    },
                    "422": {
                        "description": "Email invalide"
                    }
                }
            }
        },
        "/student/recommendations": {
            "get": {
                "tags": [
                    "Recommendations"
                ],
                "summary": "Cours recommandés selon les intérêts de l'étudiant",
                "operationId": "3f68564981e6cd7213c5613524fc3c02",
                "responses": {
                    "200": {
                        "description": "Liste des cours recommandés"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/student/interests": {
            "post": {
                "tags": [
                    "Recommendations"
                ],
                "summary": "Mettre à jour les intérêts de l'étudiant",
                "operationId": "b84155b29d85f5a1895116c81ea80568",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "interest_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2,
                                            3
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Intérêts mis à jour"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/student/wishlist": {
            "get": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Ma liste de cours sauvegardés",
                "operationId": "71c41397e8b6fc1bcfb01aefd42d7683",
                "responses": {
                    "200": {
                        "description": "Wishlist de l'étudiant"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Ajouter un cours à la wishlist",
                "operationId": "7d95ce926e90e2f7d70121c1434dd969",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "course_id"
                                ],
                                "properties": {
                                    "course_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cours ajouté à la wishlist"
                    },
                    "409": {
                        "description": "Déjà dans la wishlist"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/student/wishlist/{courseId}": {
            "delete": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Retirer un cours de la wishlist",
                "operationId": "2d9517236dfb365ae1cce98a54996e27",
                "parameters": [
                    {
                        "name": "courseId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cours retiré"
                    },
                    "404": {
                        "description": "Cours non trouvé dans la wishlist"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Recommendations",
            "description": "Recommendations"
        },
        {
            "name": "Wishlist",
            "description": "Wishlist"
        }
    ]
}